Cooling duct fans with forgotten belt stepper mount .STEP file
Left and Right Cooling duct fans. These files are very rough, but they follow the stl file pretty well. The only change is removing a side channel on the inside, so the fan is more directed on the part. Part still need fillets on most surfaces, but a redo may need to be done first. Specifically, the parametric modeling of the file is less than adequate and a lot of extra sketches/bodies. Also included is the forgotten belt stepper mount .STEP file and removing the accidental belt clamp file as well.